src/requirements.txt
streamlit
sqlalchemy 
psycopg2
python-dotenv
pdf2image
pytesseract
